Rob noticed a problem in the documentation of a function. I fixed it and expanded upon it.
We never use this function, but we have one to return a single type size value at a single breakpoint. I changed the naming convention to be more inline with our `one-element-size` mixin and made it a mixin instead of a function so you don't have to remember to add a unit declaration on the output.